[FIX] Entity type 변경

+++ woorepie/src/main/java/com/piehouse/woorepie/customer/dto/response/GetCustomerTradeResponse.java
@@ -17,9 +17,9 @@ public class GetCustomerTradeResponse {
 
     private String estateName;
 
-    private Integer tradeTokenAmount;
+    private Long tradeTokenAmount;
 
-    private Integer tradeTokenPrice;
+    private Long tradeTokenPrice;
 
     private LocalDateTime tradeDate;
